Refactor chat UI and notifications
